+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next'
 import { Noto_Sans_JP } from 'next/font/google'
 import './globals.css'
 import Header from '@/components/layouts/Header/Header'
-import NextAuthProvider from '@/utils/auth/NextAuthProvider'
+import NextAuthProvider from '@/lib/next-auth/NextAuthProvider'
 
 const notoSansJP = Noto_Sans_JP({ subsets: ['latin'] })
