+++ src/lib/next-auth/options.ts
@@ -1,0 +1,34 @@
+import { NextAuthOptions } from 'next-auth'
+import GoogleProvider from 'next-auth/providers/google'
+import { PrismaAdapter } from '@auth/prisma-adapter'
+import { PrismaClient } from '@prisma/client'
+import type { Adapter } from 'next-auth/adapters'
+
+type ClientType = {
+  clientId: string
+  clientSecret: string
+}
+
+const prisma = new PrismaClient()
+
+export const options: NextAuthOptions = {
+  debug: false,
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_CLIENT_ID ?? '',
+      clientSecret: process.env.GOOGLE_CLIENT_SECRET ?? '',
+    } as ClientType),
+  ],
+  adapter: PrismaAdapter(prisma) as Adapter,
+  callbacks: {
+    session: ({ session, user }) => {
+      return {
+        ...session,
+        user: {
+          ...session.user,
+          id: user.id,
+        },
+      }
+    },
+  },
+}
